fix(validation): address code review — t.Parallel, errors.New, package split, glob validation
- C1: add t.Parallel() to all tests and subtests
- C2: remove unused os* package-level test globals
- C3: split internal vs external test packages
- C4: use errors.New for static error messages
- I1: avoid loop-variable address in Validate
- I2: validate files_match globs at load time + test
- I3: remove dead no-conditions branch in ruleMatches
- I4: remove post-hoc sort in TestValidate
- M1: remove custom contains helper
