+++ mcp-server/internal/validation/testdata/workflow_rules/rules_and_semantics.md
@@ -1,0 +1,10 @@
+---
+rules:
+  - id: both
+    when:
+      files_match:
+        - "backend/migrations/**/*.sql"
+      title_matches: "(?i)drop"
+    require: human_gate
+    reason: "combined"
+---
