+++ mcp-server/internal/validation/testdata/workflow_rules/rules_bad_glob.md
@@ -1,0 +1,8 @@
+---
+rules:
+  - id: bad-glob
+    when:
+      files_match: ["[invalid"]
+    require: human_gate
+    reason: "invalid glob test"
+---
